+++ .github/workflows/main.yml
@@ -9,7 +9,7 @@ on:
 
 env:
   DOCKER_IMAGE: backend
-  DOCKER_TAG: 0.2
+  DOCKER_TAG: 0.3
 
 jobs:
   build:
